Reorganized app folder & added blog entities
