+++ app/Database/EntityManager.php
@@ -3,21 +3,21 @@ declare(strict_types=1);
 
 namespace App\Database;
 
-use App\Database\Entity\Customer\ApprovalPage;
-use App\Database\Entity\Localization\Language;
-use App\Database\Entity\Localization\Translation;
-use App\Database\Entity\Order\Comment;
-use App\Database\Entity\Order\Order;
-use App\Database\Entity\Order\Status;
-use App\Database\Entity\Shop;
+use App\Database\Entity\Hooray\Customer\ApprovalPage;
+use App\Database\Entity\Hooray\Localization\Language;
+use App\Database\Entity\Hooray\Localization\Translation;
+use App\Database\Entity\Hooray\Order\Comment;
+use App\Database\Entity\Hooray\Order\Order;
+use App\Database\Entity\Hooray\Order\Status;
+use App\Database\Entity\Hooray\Shop;
 use App\Database\Entity\User;
-use App\Database\Repository\Customer\ApprovalPageRepository;
-use App\Database\Repository\Localization\LanguageRepository;
-use App\Database\Repository\Localization\TranslationRepository;
-use App\Database\Repository\Order\CommentRepository;
-use App\Database\Repository\Order\OrderRepository;
-use App\Database\Repository\Order\OrderStatusRepository;
-use App\Database\Repository\ShopRepository;
+use App\Database\Repository\Hooray\Customer\ApprovalPageRepository;
+use App\Database\Repository\Hooray\Localization\LanguageRepository;
+use App\Database\Repository\Hooray\Localization\TranslationRepository;
+use App\Database\Repository\Hooray\Order\CommentRepository;
+use App\Database\Repository\Hooray\Order\OrderRepository;
+use App\Database\Repository\Hooray\Order\OrderStatusRepository;
+use App\Database\Repository\Hooray\ShopRepository;
 use App\Database\Repository\UserRepository;
 
 class EntityManager extends \Megio\Database\EntityManager
