--- app/Database/Enum/Api.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\Database\Enum;
-
-enum Api: string
-{
-    case WOOCOMMERCE = 'woocommerce';
-}
