+++ app/Database/Enum/Hooray/Api.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Database\Enum\Hooray;
+
+enum Api: string
+{
+    case WOOCOMMERCE = 'woocommerce';
+}
